apps/agent/pyproject.toml
[project]
name = "baserank-agent"
version = "1.0.0"
description = "BaseRank Protocol AI Agent for autonomous score updates"
requires-python = ">=3.10"
dependencies = [
    "coinbase-agentkit>=0.1.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "schedule>=1.2.0",
    "httpx>=0.27.0",
]

[project.scripts]
agent = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
